fix: :zap: TcpSender connection and TcpListener Accept
Improved the way in which the sender connect to the server, an the listener accept incoming connections. In particular now the sender has the possibility to retry unsuccesful connection and the listener has a timeout on the accepting to be able to stop the listening thread whenever we'd like to. Finished the Communication Test.