feat: :sparkles: Introduced the Communication Interface.
The communication interface is a sent-n-receive entity for network communication able to send and listen for incoming messages. TcpCommunicationInterface and UdpCommunicationInterface are the subclasses implementing UdpListener and UdpSender, and TcpListener and TcpSender. Finally, some new fixed have been done to Thread class introducing isRunning method. First idea of configuration folder