feat: ✨ Add Discover with ProgBar. Lua Dissectors for Wireshark and Modified conf
Finished the sending of DISCOVER HELLO messages using the discover protocol and starting the implmentation of the reception queue in the QubeInterface. Modified the configuration file adding the default UDP listening port for qube worker. Modified the DMD CommonHeader adding also the Protocol FLAG field. Added wireshark dissector for the common header and the discover hello protocol.